+++ app/src/main/java/reikai/domain/novel/NovelTrackRepository.kt
@@ -1,0 +1,17 @@
+package reikai.domain.novel
+
+import kotlinx.coroutines.flow.Flow
+import reikai.domain.novel.model.NovelTrack
+
+interface NovelTrackRepository {
+
+    suspend fun getTrackById(id: Long): NovelTrack?
+
+    suspend fun getTracksByNovelId(novelId: Long): List<NovelTrack>
+
+    fun getTracksByNovelIdAsFlow(novelId: Long): Flow<List<NovelTrack>>
+
+    suspend fun delete(novelId: Long, trackerId: Long)
+
+    suspend fun insert(track: NovelTrack)
+}